Move handlers for hotkeys into the deliveries ul
